+++ main_ml
@@ -1,0 +1,42 @@
+import pandas as pd
+import argparse
+import torch
+import torch.nn as nn
+from sklearn.model_selection import train_test_split
+from torchvision import models as models
+from torch.utils.data import DataLoader
+from models.SVM import *
+from models.KNN import *
+from models.DT import *
+from models.ExtraTrees import *
+from models.RandomForest import *
+from models.Logistic import *
+from models.GradientBoosting import *
+import os
+import argparse
+from utils import *
+from train import *
+from model_lstm import *
+
+path_dir = "./data/Preprocessed"
+file_list = os.listdir(path_dir)
+pkl_list = []
+for filename in file_list:
+    pkl_list.append('./data/Preprocessed/'+filename)
+    
+train_dataPaths, test_dataPaths = train_test_split(pkl_list, train_size = 0.8, random_state = 1,shuffle=True)
+
+(X_trains, y_trains) = preprocess_from_pklpaths(train_dataPaths)
+(X_tests, y_tests)   = preprocess_from_pklpaths(test_dataPaths)
+
+X_train = pd.concat(X_trains)
+y_train = pd.concat(y_trains)
+X_test  = pd.concat(X_tests)
+y_test  = pd.concat(y_tests)
+
+Decision_Tree(X_train,X_test,y_train,y_test)
+SVM(X_train,X_test,y_train,y_test)
+KNN(X_train,X_test,y_train,y_test)
+RandomForest(X_train,X_test,y_train,y_test)
+ExtraTrees(X_train,X_test,y_train,y_test)
+GradientBoosting(X_train,X_test,y_train,y_test)
